Prediction of Housing Prices Based Using ML
Evaluation and Prediction of Housing Prices Based on Building Features and Geographical Location Using Machine Learning  
In this project, housing prices in Beijing, China, were evaluated and predicted based on house features and related information such as elevation, number of floors, renovation status, distance from the city center, and house area. The predicted prices were then visualized on a real map, with a heatmap displaying housing prices in different regions.  
In the final stage, housing prices were predicted over time based on temporal data and regional analysis, and time series charts were created to illustrate the trends.  
This project enables the estimation of future housing prices by analyzing the real map for different areas in Beijing. Additionally, it provides a basis for urban planning and development of construction projects in the city.